Refactoring code and updating ReadMe file

+++ app/src/main/java/com/example/androiddevelopercodechallenge/data/dataSource/local/LocalDataSourceImpl.kt
@@ -4,6 +4,7 @@ import android.util.Log
 import androidx.paging.PagingSource
 import com.example.androiddevelopercodechallenge.data.model.Result
 import com.example.androiddevelopercodechallenge.data.roomDB.ResultDAO
+import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
@@ -22,6 +23,9 @@ class LocalDataSourceImpl @Inject constructor(
                 resultDAO.insertAllResults(results = results)
                 Log.d("MyTag", "LocalDataSourceImpl: insertAllResults(): successfully completed")
             } catch (e: Exception) {
+                if (e is CancellationException) {
+                    throw e
+                }
                 Log.e("MyTag", "LocalDataSourceImpl: insertAllResults(): failed ${e.message}")
             }
         }
@@ -32,10 +36,13 @@ class LocalDataSourceImpl @Inject constructor(
     }
 
     override fun getAllResults(): Flow<List<Result>> {
-       return try {
+        return try {
             Log.d("MyTag", "LocalDataSourceImpl: getAllResults(): successfully completed")
             resultDAO.getAllResults()
         } catch (e: Exception) {
+            if (e is CancellationException) {
+                throw e
+            }
             Log.e("MyTag", "LocalDataSourceImpl: getAllResults(): failed ${e.message}")
             emptyFlow()
         }
@@ -46,6 +53,9 @@ class LocalDataSourceImpl @Inject constructor(
             resultDAO.addResult(result = result)
             Log.d("MyTag", "LocalDataSourceImpl: addResult(): successfully completed")
         } catch (e: Exception) {
+            if (e is CancellationException) {
+                throw e
+            }
             Log.e("MyTag", "LocalDataSourceImpl: addResult(): failed ${e.message}")
         }
     }
@@ -55,6 +65,9 @@ class LocalDataSourceImpl @Inject constructor(
             resultDAO.updateResult(result = result)
             Log.d("MyTag", "LocalDataSourceImpl: updateResult(): successfully completed")
         } catch (e: Exception) {
+            if (e is CancellationException) {
+                throw e
+            }
             Log.e("MyTag", "LocalDataSourceImpl: updateResult(): failed ${e.message}")
         }
     }
@@ -68,16 +81,31 @@ class LocalDataSourceImpl @Inject constructor(
                     "LocalDataSourceImpl: deleteResultsById(): successfully completed: email:$email"
                 )
             } catch (e: Exception) {
+                if (e is CancellationException) {
+                    throw e
+                }
                 Log.e("MyTag", "LocalDataSourceImpl: deleteResultsById(): failed ${e.message}")
             }
         }
 
-    override suspend fun deleteAll() {
+    override suspend fun deleteAll(): Unit = withContext(coroutineDispatcher) {
         try {
             resultDAO.deleteAll()
             Log.d("MyTag", "LocalDataSourceImpl: deleteAll(): successfully completed")
         } catch (e: Exception) {
+            if (e is CancellationException) {
+                throw e
+            }
             Log.e("MyTag", "LocalDataSourceImpl: deleteAll(): failed ${e.message}")
+        }
+    }
+
+    override suspend fun getResultsCount(): Int = withContext(coroutineDispatcher) {
+        return@withContext try {
+            resultDAO.getResultsCount()
+        } catch (e: Exception) {
+            Log.e("MyTag","LocalDataSourceImpl: getResultsCount: failed :${e.message}")
+            0
         }
     }
 }
